MisinfoGuard/Backend/requirements.txt
google-generativeai
groq
duckduckgo-search
python-dotenv
rich
fastapi
uvicorn
pydantic
PyJWT
passlib
bcrypt
email-validator
